Consolidate internal::RecordReader, ColumnReader files
